requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cssselect2==0.9.0
deep-translator==1.11.4
fastapi==0.135.1
fastapi-limiter==0.1.6
fonttools==4.62.0
h11==0.16.0
httptools==0.7.1
idna==3.11
iniconfig==2.3.0
lxml==6.0.2
packaging==26.0
passlib==1.7.4
pillow==12.1.1
pluggy==1.6.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydyf==0.12.1
Pygments==2.19.2
PyJWT==2.11.0
pyphen==0.17.2
pyrate-limiter==4.0.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
redis==7.3.0
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
tinycss2==1.5.1
tinyhtml5==2.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
weasyprint==68.1
webencodings==0.5.1
websockets==16.0
zopfli==0.4.1

# --- ML / AI Module Dependencies ---
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
shap>=0.43.0
imbalanced-learn>=0.11.0
joblib>=1.3.0

# --- MLOps ---
mlflow>=2.8.0
evidently>=0.4.0
